refactor(app): move the setup menu state into ainb-app
SetupMenuItem and SetupMenuState move to ainb_app::components::setup_menu;
the menu renderer stays in ainb-core.
